// src/roles/enums/user-role.enum.ts
export enum UserRole {
  SUPER_USER = 'SUPER_USER',
  ACCOUNT_ADMINISTRATOR = 'ACCOUNT_ADMINISTRATOR',
  BILLING_ADMINISTRATOR = 'BILLING_ADMINISTRATOR',
  CALL_QUEUE_ADMINISTRATOR = 'CALL_QUEUE_ADMINISTRATOR',
  CALL_QUEUE_REPORTING_ADMINISTRATOR = 'CALL_QUEUE_REPORTING_ADMINISTRATOR',
  REPORTS_ADMINISTRATOR = 'REPORTS_ADMINISTRATOR',
  USER_ADMINISTRATOR = 'USER_ADMINISTRATOR',
  END_USER = 'END_USER',
  END_USER_NO_DASHBOARD = 'END_USER_NO_DASHBOARD',
}